Move public API to new location
