+++ src/PowerShellEditorServices/Extensions/EditorCommand.cs
@@ -5,7 +5,7 @@
 
 using System.Management.Automation;
 
-namespace Microsoft.PowerShell.EditorServices.Services.PowerShellContext
+namespace Microsoft.PowerShell.EditorServices.Extensions
 {
     /// <summary>
     /// Provides details about a command that has been registered
