+++ src/PowerShellEditorServices/Extensions/EditorTerminal.cs
@@ -3,13 +3,13 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-namespace Microsoft.PowerShell.EditorServices.Services.PowerShellContext
+namespace Microsoft.PowerShell.EditorServices.Extensions
 {
     /// <summary>
     /// Provides a PowerShell-facing API which allows scripts to
     /// interact with the editor's terminal.
     /// </summary>
-    public sealed class EditorTerminal
+    public class EditorTerminal
     {
         #region Private Fields
